requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
fasttext-wheel==0.9.2
